refactor monitoring command router
